% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{baixar_tabela_stf}
\alias{baixar_tabela_stf}
\title{Baixar tabelas STF}
\usage{
baixar_tabela_stf(naMarra = FALSE)
}
\arguments{
\item{naMarra}{se TRUE, sobrescreve arquivo baixado. Por padrão, é FALSE.}
}
\value{
caminho para o arquivo baixado
}
\description{
Baixar tabelas STF
}
