% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{ler_tabela_stf}
\alias{ler_tabela_stf}
\title{Ler tabela STF}
\usage{
ler_tabela_stf(data = "mais_recente")
}
\arguments{
\item{data}{data em formato AAAA-MM-DD. Por padrão, é "mais_recente".}
}
\value{
tibble com processos distribuídos
}
\description{
Ler tabela STF
}
